==> Big data analysis/Semester 1/02_pandas_.ipynb <==
# Pandas

Материалы:
* Макрушин С.В. "Лекция 2: Библиотека Pandas"
* https://pandas.pydata.org/docs/user_guide/index.html#
* https://pandas.pydata.org/docs/reference/index.html
* Уэс Маккини. Python и анализ данных

## Задачи для совместного разбора

1. Загрузите данные из файла `sp500hst.txt` и обозначьте столбцы в соответствии с содержимым: `"date", "ticker", "open", "high", "low", "close", "volume"`.

import pandas as pd
df=pd.read_csv('/content/sp500hst.txt', header=None, names=["date", "ticker", 'open','high','low','close','volume'])
print (df)
pd.to_datetime(df['date'], format='%Y%m%d').dt.month

df['ticker'].str.lower()

2. Рассчитайте среднее значение показателей для каждого из столбцов c номерами 3-6.

3. Добавьте столбец, содержащий только число месяца, к которому относится дата.

4. Рассчитайте суммарный объем торгов для для одинаковых значений тикеров.

5. Загрузите данные из файла sp500hst.txt и обозначьте столбцы в соответствии с содержимым: "date", "ticker", "open", "high", "low", "close", "volume". Добавьте столбец с расшифровкой названия тикера, используя данные из файла `sp_data2.csv` . В случае нехватки данных об именах тикеров корректно обработать их.

sp = pd.read_csv('/content/sp_data2.csv', sep =';', names=['ticker', 'company', 'percent'], header=None)
sp.head()

pd.merge(df,sp,how="inner", left_on="ticker", right_on="ticker")

## Лабораторная работа №2

### Базовые операции с `DataFrame`

1.1 В файлах `recipes_sample.csv` и `reviews_sample.csv` находится информация об рецептах блюд и отзывах на эти рецепты соответственно. Загрузите данные из файлов в виде `pd.DataFrame` с названиями `recipes` и `reviews`. Обратите внимание на корректное считывание столбца(ов) с индексами.

recipes  = pd.read_csv('/content/recipes_sample.csv', sep =',', names=["name","id","minutes","contributor_id","submitted","n_steps","description","n_ingredients"], header=None, skiprows=1)
reviews = pd.read_csv('/content/reviews_sample.csv', sep =',', names=["ID","user_id","recipe_id","date","rating","review"], header=None , skiprows=1)
recipes


1.2 Для каждой из таблиц выведите основные параметры:
* количество точек данных (строк);
* количество столбцов;
* тип данных каждого столбца.

index_recipes = recipes.index
index_reviews = reviews.index
print (len (index_recipes), "number of recipes\n", len(index_reviews),"number of reviews")
print(len(recipes.columns)," num of columns in recipes")
print(len(reviews.columns), "num of columns in reviews")
print (recipes.dtypes, "types of recipies\n\n", reviews.dtypes,"types of reviews")

1.3 Исследуйте, в каких столбцах таблиц содержатся пропуски. Посчитайте долю строк, содержащих пропуски, в отношении к общему количеству строк.

print ("Number of nan's in recipes \n",recipes.isna().sum().sum(),"\nNumber of nan's in reviews\n", reviews.isna().sum().sum())
print ("\n",recipes.isnull().mean() * 100)
print ("\n",reviews.isnull().mean() * 100)

1.4 Рассчитайте среднее значение для каждого из числовых столбцов (где это имеет смысл).

recipes['minutes'] = pd.to_numeric(recipes['minutes'])
recipes['n_steps'] = pd.to_numeric(recipes['n_steps'])
recipes['n_ingredients'] = pd.to_numeric(recipes['n_ingredients'])
recipes['submitted']= pd.to_datetime(recipes['submitted'])
print ("Recipes\n",recipes['minutes'].mean(), "minutes\n",recipes['n_steps'].mean(), "n_steps\n", recipes['n_ingredients'].mean(), "n_ingredients" )

1.5 Создайте серию из 10 случайных названий рецептов.

repa = recipes['name'].sample(n=10)
print (repa)

1.6 Измените индекс в таблице `reviews`, пронумеровав строки, начиная с нуля.

repa=reviews
repa['ID']=range(0, len(repa))
print (repa)


1.7 Выведите информацию о рецептах, время выполнения которых не больше 20 минут и кол-во ингредиентов в которых не больше 5.

recipes  = pd.read_csv('/content/recipes_sample.csv', sep =',', names=["name","id","minutes","contributor_id","submitted","n_steps","description","n_ingredients"], header=None, skiprows=1)

temp = recipes[(recipes['minutes'] < 20) & (recipes['n_ingredients'] <= 5)]
temp



### Работа с датами в `pandas`

2.1 Преобразуйте столбец `submitted` из таблицы `recipes` в формат времени. Модифицируйте решение задачи 1.1 так, чтобы считать столбец сразу в нужном формате.

recipes  = pd.read_csv('/content/recipes_sample.csv', sep =',', names=["name","id","minutes","contributor_id","submitted","n_steps","description","n_ingredients"], header=None, skiprows=1)
recipes['submitted']= pd.to_datetime(recipes['submitted'])


2.2 Выведите информацию о рецептах, добавленных в датасет не позже 2010 года.

temp=(recipes['submitted']<"2011-01-01")
print (recipes[temp])

### Работа со строковыми данными в `pandas`

3.1  Добавьте в таблицу `recipes` столбец `description_length`, в котором хранится длина описания рецепта из столбца `description`.

recipes  = pd.read_csv('/content/recipes_sample.csv', sep =',', names=["name","id","minutes","contributor_id","submitted","n_steps","description","n_ingredients"], header=None, skiprows=1)

recipes['description_length']  = recipes['description'].str.len()
recipes


3.2 Измените название каждого рецепта в таблице `recipes` таким образом, чтобы каждое слово в названии начиналось с прописной буквы.

recipes['name']  = recipes['name'].str.title()
recipes

3.3 Добавьте в таблицу `recipes` столбец `name_word_count`, в котором хранится количество слов из названии рецепта (считайте, что слова в названии разделяются только пробелами).

recipes['name_word_count']  = recipes['name'].str.split().str.len()
recipes

### Группировки таблиц `pd.DataFrame`

4.1 Посчитайте количество рецептов, представленных каждым из участников (`contributor_id`). Какой участник добавил максимальное кол-во рецептов?

recipes.groupby('id')['id'].count().idxmax()

4.2 Посчитайте средний рейтинг к каждому из рецептов. Для скольких рецептов отсутствуют отзывы?

reviews = pd.read_csv('/content/reviews_sample.csv', sep =',', names=["ID","user_id","recipe_id","date","rating","review"], header=None , skiprows=1)
print(reviews.groupby('recipe_id')['rating'].mean())
print ("\nnum of nans")
reviews['review'].isna().sum()


4.3 Посчитайте количество рецептов с разбивкой по годам создания.

reviews.index=pd.to_datetime(reviews['date'])
print(reviews.groupby(pd.Grouper(freq='Y')).size())

### Объединение таблиц `pd.DataFrame`

5.1 При помощи объединения таблиц, создайте `DataFrame`, состоящий из четырех столбцов: `id`, `name`, `user_id`, `rating`. Рецепты без отзывов должны отсутствовать в данной таблице. Подтвердите правильность работы вашего кода, выбрав рецепт, не имеющий отзывов, и выведя на экран строку из полученного `DataFrame`, содержащую информацию об этом отзыве.

reviews = pd.read_csv('/content/reviews_sample.csv', sep =',', names=["id","user_id","recipe_id","date","rating","review"], header=None , skiprows=1)
recipes  = pd.read_csv('/content/recipes_sample.csv', sep =',', names=["name","id","minutes","user_id","submitted","n_steps","description","n_ingredients"], header=None, skiprows=1)
reviews = reviews[reviews['review'].notna()]
reviews.pop("recipe_id")
reviews.pop("date")
reviews.pop("review")
reviews.join( recipes.set_index('id')[['name']], on='user_id')


5.2 При помощи объединения таблиц и группировок, создайте `DataFrame`, состоящий из трех столбцов: `recipe_id`, `name`, `review_count`. У рецептов, для которых отсутствуют отзывы, в соответствущем столбце должен быть указан 0. Подтвердите правильность работы вашего кода, выбрав рецепт, не имеющий отзывов, и выведя на экран строку из полученного `DataFrame`, содержащую информацию об этом отзыве.


reviews = pd.read_csv('/content/reviews_sample.csv', sep =',', names=["id","user_id","recipe_id","date","rating","review"], header=None , skiprows=1)
recipes  = pd.read_csv('/content/recipes_sample.csv', sep =',', names=["name","id","minutes","user_id","submitted","n_steps","description","n_ingredients"], header=None, skiprows=1)
recipes=recipes.join( reviews.set_index('recipe_id')[['review']], on='id')
recipes.pop("n_steps")
recipes.pop("minutes")
recipes.pop("submitted")
recipes.pop("description")
recipes.pop("user_id")
recipes.pop("n_ingredients")
#recipes.groupby('id').review.apply(lambda x: x.count()-x.isna().sum())
recipes.review.notnull().groupby([recipes['name'],recipes['id']]).sum().astype(int).reset_index(name='count')


5.3. Выясните, отзывы, добавленные в каком году, имеют наименьший средний рейтинг?

reviews = pd.read_csv('/content/reviews_sample.csv', sep =',', names=["id","user_id","recipe_id","date","rating","review"], header=None , skiprows=1)
recipes  = pd.read_csv('/content/recipes_sample.csv', sep =',', names=["name","id","minutes","user_id","submitted","n_steps","description","n_ingredients"], header=None, skiprows=1)

reviews.index=pd.to_datetime(reviews['date'])
#dfc = reviews.groupby(pd.Grouper(freq='Y'))['rating']
reviews=reviews.groupby(pd.Grouper(freq='Y'))['rating'].mean().nsmallest(1)
reviews

### Сохранение таблиц `pd.DataFrame`

6.1 Отсортируйте таблицу в порядке убывания величины столбца `name_word_count` и сохраните результаты выполнения заданий 3.1-3.3 в csv файл. 

recipes  = pd.read_csv('/content/recipes_sample.csv', sep =',', names=["name","id","minutes","contributor_id","submitted","n_steps","description","n_ingredients"], header=None, skiprows=1)

recipes['description_length']  = recipes['description'].str.len()
recipes['name']  = recipes['name'].str.capitalize()
recipes['name_word_count']  = recipes['name'].str.split().str.len()
recipes.to_csv('raw_data.csv', index=False)

6.2 Воспользовавшись `pd.ExcelWriter`, cохраните результаты 5.1 и 5.2 в файл: на лист с названием `Рецепты с оценками` сохраните результаты выполнения 5.1; на лист с названием `Количество отзывов по рецептам` сохраните результаты выполнения 5.2.

reviews = pd.read_csv('/content/reviews_sample.csv', sep =',', names=["id","user_id","recipe_id","date","rating","review"], header=None , skiprows=1)
recipes  = pd.read_csv('/content/recipes_sample.csv', sep =',', names=["name","id","minutes","user_id","submitted","n_steps","description","n_ingredients"], header=None, skiprows=1)
reviews = reviews[reviews['review'].notna()]
reviews.pop("recipe_id")
reviews.pop("date")
reviews.pop("review")
res1=reviews.join( recipes.set_index('id')[['name']], on='user_id')

with pd.ExcelWriter("Рецепты с оценками.xlsx") as writer:
  res1.to_excel(writer)

reviews = pd.read_csv('/content/reviews_sample.csv', sep =',', names=["id","user_id","recipe_id","date","rating","review"], header=None , skiprows=1)
recipes  = pd.read_csv('/content/recipes_sample.csv', sep =',', names=["name","id","minutes","user_id","submitted","n_steps","description","n_ingredients"], header=None, skiprows=1)
recipes=recipes.join( reviews.set_index('recipe_id')[['review']], on='id')
recipes.pop("n_steps")
recipes.pop("minutes")
recipes.pop("submitted")
recipes.pop("description")
recipes.pop("user_id")
recipes.pop("n_ingredients")
#recipes.groupby('id').review.apply(lambda x: x.count()-x.isna().sum())
res2=recipes.review.notnull().groupby([recipes['name'],recipes['id']]).sum().astype(int).reset_index(name='count')

with pd.ExcelWriter("Количество отзывов по рецептам.xlsx") as writer:
  res2.to_excel(writer)
