==> Big data analysis/Semester 1/05_data_files_2_Фамилия_Имя.ipynb <==
# Форматы данных (2)

from google.colab import drive
drive.mount('/content/drive')

Материалы:
* Макрушин С.В. "Лекция 5: Форматы данных (часть 2)"
* https://docs.python.org/3/library/csv.html
* https://docs.h5py.org/en/stable/
* Уэс Маккини. Python и анализ данных

## Задачи для совместного разбора

1. Считайте данные из файла `open_pubs.csv`, используя `csv.reader`, и преобразуйте к структуре данных следующего вида:
    
`{'fas_id': [24, 30, ...], 'name': ['Achor Inn', 'Angel Inn', ...], ... }`

import csv
 
with open('/content/drive/MyDrive/Colab Notebooks/lab 5/open_pubs.csv') as fp:
    reader = csv.reader(fp)
    header = next(reader)
    
    for row in reader:
        print(row)
        break

2. Сгенерируйте 2 случайные матрицы размера 10_000 x 10_000 и вычислите их произведение. Сколько времени занимают три этих операции? Сохраните 3 полученных матрицы в файл .npz с соответствующими названиями

import numpy as np
 
A = np.random.randint(0, 100, size=(10_000, 10_000))
B = np.random.randint(0, 100, size=(10_000, 10_000))
 
np.save('A.npy', A)
np.savez('AB.npz', A, B)
 
r = np.load('AB.npz')
print(r.files)
 
r['arr_0']
 
A.dtype
 
10_000 * 10_000 * 4 / 1024 / 1024

3. Создайте 2 матрицы размера 1000x1000, используя различные параметризируемые распределения из numpy (https://docs.scipy.org/doc/numpy-1.15.0/reference/routines.random.html#distributions)

После этого сохраните получившиеся матрицы в hdf5-файл в виде двух различных датасетов. В качестве описания каждого датасета укажите параметры используемых распределений 

import h5py
 
with h5py.File('test.h5', 'w') as hdf:
    ds1 = hdf.create_dataset('arrA', data=A)
    ds2 = hdf.create_dataset('arrB', data=B)
    
    ds1.attrs['description'] = 'Здесь лежит массив А'
 
with h5py.File('test.h5', 'r') as hdf:
    ds1 = hdf['arrA']
    print(type(ds1))
    arr = ds1[:100]
# ..
 
ds1[:100]

## Лабораторная работа 5

### csv

1.1 В файле `tags_sample.csv` находится информация о тэгах, приписываемых рецептам. Воспользовавшись `csv.reader`, считайте этот файл и создайте словарь вида `id_рецепта: [список тэгов]`. Сохраните этот словарь в файл `tags_sample.json`.

import csv
import json


data_dict = {}

with open('/content/drive/MyDrive/Colab Notebooks/lab 5/tags_sample.csv') as file:
    reader = csv.reader(file)
    header = next(reader)
    print(header)
    for row in reader:
        
        item_id = int(row[0])
        item_tag = row[1]
        
        if item_id not in data_dict:
            data_dict[item_id] = []
        data_dict[item_id].append(item_tag)

1.2 Считайте файл `recipes_sample_with_filled_nsteps.csv` (__ЛР4__) в виде `pd.DataFrame`. Добавьте к таблице 2 столбца: `n_tags`, содержащий количество тэгов у этого рецепта; и `tags`, содержащий набор тэгов в виде строки (тэги внутри строки разделяются символом `;`)

with open("/content/result.json", "w") as file:
    json.dump(data_dict, file)

with open("/content/result.json", "r") as file:
    data_dict_new = json.load(file)



import pandas as pd

recipes_filled = pd.read_csv("/content/drive/MyDrive/Colab Notebooks/lab 4/recipes_sample_with_filled_nsteps.csv")
#print(recipes_filled)


patched_dict = {k : len(v) for (k,v) in data_dict.items()}

def filler(x):
    if x["id"] in patched_dict:
        return patched_dict[x["id"]] 
    return float("nan")
    
recipes_filled['n_tags'] = recipes_filled.apply(lambda x: filler(x), axis=1)


print(recipes_filled)

patched_dict = {k : ";".join(v) for (k,v) in data_dict.items()}

print(patched_dict)



def filler(x):
    if x["id"] in patched_dict:
        return patched_dict[x["id"]]
    return float("nan")

recipes_filled['tags'] = recipes_filled.apply(lambda x: filler(x), axis=1)


print(recipes_filled)

print (recipes_filled[recipes_filled['n_tags'].isna()] )


1.3 В файле `ingredients_sample.csv` находится информация о ингредиентах, необходимых для рецепта. Воспользовавшись `csv.DictReader`, считайте этот файл и создайте словарь вида `id_рецепта: [список ингредиентов]`.

data_dict = {}

with open('/content/drive/MyDrive/Colab Notebooks/lab 5/ingredients_sample.csv') as file:
    reader = csv.DictReader(file)
    for item in reader:
        ingredient, recipe_id = item["ingredient"], int(item["recipe_id"])

        if recipe_id not in data_dict:
            data_dict[recipe_id] = []
        data_dict[recipe_id].append(ingredient)

        
data_dict



1.4 Добавьте к таблице из задания 1.2 столбец `ingredients`, содержащий набор ингредиентов в виде строки (ингредиенты внутри строки разделяются символом `*`)

Для строк, которые содержат пропуски в столбце `n_ingredients`, заполните их на основе файла  `ingredients_sample.csv`

import pandas as pd

patched_dict = {k : "*".join(v) for k, v in data_dict.items()}

def filler(x):
    if x["id"] in patched_dict:
        return patched_dict[x["id"]]
    return float("nan")
    
recipes_filled['ingredients'] = recipes_filled.apply(lambda x: filler(x), axis=1)


print (recipes_filled)


patched_dict = {k : len(v) for k, v in data_dict.items()}


def filler(x):
    if pd.isna(x["n_ingredients"]) and x["id"] in patched_dict:
        return patched_dict[x["id"]]
    return x["n_ingredients"]

recipes_filled['n_ingredients'] = recipes_filled.apply(lambda x: filler(x), axis=1)


print (recipes_filled)



1.5 Проверьте, содержит ли столбец `n_ingredients` пропуски. Если нет, треобразуйте его к целочисленному типу и сохраните результаты в файл `recipes_sample_with_tags_ingredients.csv`

recipes_filled[recipes_filled["n_ingredients"].isnull()]


recipes_filled["n_ingredients"] = recipes_filled["n_ingredients"].astype(int)


print(recipes_filled.dtypes)

print (recipes_filled)

recipes_filled.to_csv("./recipes_sample_with_tags_ingredients.csv")

### npy

2.1 Разделите таблицу, полученную в результате 1.5, на две таблицы: одна содержит рецепты, загруженные до 2000 года; вторая - все остальные. В полученных таблицах оставьте только числовые столбцы и преобразуйте их к `numpy.array`

import pandas as pd

recipes_filled["submitted"] = pd.to_datetime(recipes_filled["submitted"], format='%Y-%m-%d', errors='ignore')

print(recipes_filled)

print (recipes_filled.dtypes)



recipes_before = recipes_filled[recipes_filled["submitted"].dt.year < 2000]
print (recipes_before)

recipes_after = recipes_filled[recipes_filled["submitted"].dt.year >= 2000]
print (recipes_after)

print (recipes_after.dtypes)


recipes_after_np = np.array([recipes_after[item] for item in ["id", "minutes", "contributor_id","n_steps","n_ingredients","n_tags"]]) # recipes_after[["id", "minutes"]].values

print (recipes_after_np)

recipes_before_np = np.array([recipes_before[item] for item in ["id", "minutes", "contributor_id","n_steps","n_ingredients","n_tags"]])
print (recipes_before_np)



2.2. Сохраните 2 полученных массива в архив `npz`. Дайте массивам читаемые имена.

import h5py
import numpy as np

my_path = "./task_2_2.npz"
np.savez(my_path, recipes_before_2000=recipes_before_np, recipes_after_2000=recipes_after_np )

2.3 Считайте созданный архив и продемонстрируйте, что данные считались корректно. 



data = np.load(my_path)
data.files

recipes_before_2000 = data["recipes_before_2000"]
print (recipes_before_2000)



recipes_after_2000 = data["recipes_after_2000"]
print(recipes_after_2000)



### hdf

3.1 Выведите названия всех датасетов, находящихся в файле `nutrition_sample.h5`, а также размерность матриц, содержащихся в данных датасетах и их метаданные.

Формат вывода:
```
Dataset name=dataset_0, dataset size=(30000,), metadata={'info': 'calories (#)'}
Dataset name=dataset_1, dataset size=(30000,), metadata={'info': 'total fat (PDV)'}
...
```

datasets_dict = {}
with h5py.File("/content/drive/MyDrive/Colab Notebooks/lab 5/nutrition_sample.h5","r") as hdf:
    
    for key in hdf.keys():
        file = hdf[key]
        metadata = dict(file.attrs)
        print(f"Dataset name={key}, dataset size={file.shape}, metadata={metadata}")
        datasets_dict[key] = {"all" : {'data': file[:]}, 'attrs': metadata}



datasets_dict



3.2 Разбейте каждый из имеющихся датасетов на две части: 1 часть содержит только те строки, где PDV (Percent Daily Value) превышает 100%; 2 часть содержит те строки, где PDV не составляет не более 100%. Создайте 2 группы в файле и разместите в них соответствующие части датасета c сохранением метаданных исходных датасетов. Итого должно получиться 2 группы, содержащие несколько датасетов. Сохраните результаты в файл `nutrition_grouped.h5`

datasets_dict = {k : v for k, v in datasets_dict.items() if "PDV" in v["attrs"]["col_1"]}

new_datasets_dict = {"PDV_more_100" : {}, "PDV_less_100": {}}
for dataset, values in datasets_dict.items():

    data = values["all"]["data"]
    mask = data[:,1] > 1
    
    new_datasets_dict["PDV_more_100"][dataset] = {"data" :  data[mask], "attrs" : values["attrs"], "name": dataset}

    mask = data[:,1] < 1
    new_datasets_dict["PDV_less_100"][dataset] = {"data" :  data[mask], "attrs" : values["attrs"], "name": dataset}


for dataset, data in new_datasets_dict.items():
    print(dataset)
    print(data)




h5_path = "./nutrition_grouped.h5"
with h5py.File(h5_path, mode='w') as hdf:
    for group, data in new_datasets_dict.items():
        new_group = hdf.create_group(group)
        for d in data:
            item = new_group.create_dataset(name=data[d]['name'], data=data[d]['data'])
            item.attrs.update(data[d]['attrs'])



3.3 Выведите названия всех групп и датасетов, находящихся в этих группах, из файла `nutrition_grouped.h5` а также размерность матриц, содержащихся в датасетах и их метаданные.

with h5py.File(h5_path) as hdf:
    
    for group_name, data in hdf.items():
        print(f'\nGroup: {group_name}')

        for key in data:
            file = data[key]
            metadata = dict(file.attrs)
            print(f"Dataset name={key}, dataset size={file.shape}, metadata={metadata}")
            datasets_dict[key] = {"all" : {'data': file[:]}, 'attrs': metadata}

3.4 Модифицируйте код из 3.3 таким образом, чтобы сохранить датасеты, используя сжатие. Сравните размер полученного файла с размерами файла из 3.3. Прокомментируйте результат.

h5_compressed_path = "./nutrition_grouped_compressed.h5"
with h5py.File(h5_compressed_path, mode='w') as hdf:
    for group, data in new_datasets_dict.items():
        new_group = hdf.create_group(group)
        for d in data:
            item = new_group.create_dataset(name=data[d]['name'], data=data[d]['data'], compression="gzip", compression_opts=9)
            item.attrs.update(data[d]['attrs'])


def get_file_size_2(file):
 
    stat = os.stat(file)
    print(stat)
    size = stat.st_size
    return size

def convert_bytes(size, unit=None):
    if unit == "KB":
        return print('File size: ' + str(round(size / 1024, 3)) + ' Kilobytes')
    elif unit == "MB":
        return print('File size: ' + str(round(size / (1024 * 1024), 3)) + ' Megabytes')
    elif unit == "GB":
        return print('File size: ' + str(round(size / (1024 * 1024 * 1024), 3)) + ' Gigabytes')
    else:
        return print('File size: ' + str(size) + ' bytes')


import os


size = get_file_size_2(h5_path)
print('File size: ' + str(size) + ' bytes')

size = get_file_size_2(h5_compressed_path)
print('File size: ' + str(size) + ' bytes')
