==> Big data analysis/Semester 1/09_string_2_Фамилия_Имя.ipynb <==
# Введение в обработку текста на естественном языке

Материалы:
* Макрушин С.В. Лекция 9: Введение в обработку текста на естественном языке\
* https://realpython.com/nltk-nlp-python/
* https://scikit-learn.org/stable/modules/feature_extraction.html

## Задачи для совместного разбора

from sklearn.feature_extraction.text import CountVectorizer
import pymorphy2

1. Считайте слова из файла `litw-win.txt` и запишите их в список `words`. В заданном предложении исправьте все опечатки, заменив слова с опечатками на ближайшие (в смысле расстояния Левенштейна) к ним слова из списка `words`. Считайте, что в слове есть опечатка, если данное слово не содержится в списке `words`. 

text = '''с велечайшим усилием выбравшись из потока убегающих людей Кутузов со свитой уменьшевшейся вдвое поехал на звуки выстрелов русских орудий'''

2. Разбейте текст из формулировки задания 1 на слова; проведите стемминг и лемматизацию слов.

3. Преобразуйте предложения из формулировки задания 1 в векторы при помощи `CountVectorizer`.

## Лабораторная работа 9

### Расстояние редактирования

from google.colab import drive
drive.mount('/content/drive')

1.1 Загрузите предобработанные описания рецептов из файла `preprocessed_descriptions.csv`. Получите набор уникальных слов `words`, содержащихся в текстах описаний рецептов (воспользуйтесь `word_tokenize` из `nltk`). 

import nltk
nltk.download('punkt')

import pandas as pd
from nltk.tokenize import word_tokenize
import random

preprocessed = pd.read_csv("/content/drive/MyDrive/Colab Notebooks/lab 9/preprocessed_descriptions.csv")

def task_1_1(words):

  for index, row in preprocessed.iterrows():
      try:
          for word in word_tokenize(row[1]):
              words.add(word)
      except TypeError:
          pass
  return (words)

words = set()
task_1_1(words)
words

1.2 Сгенерируйте 5 пар случайно выбранных слов и посчитайте между ними расстояние редактирования.

from nltk.metrics.distance import edit_distance

group1 = random.sample(words, 5)
group2 = random.sample(words, 5)
for i in range(len(group1)):
    print(edit_distance(group1[i], group2[i]))

1.3 Напишите функцию, которая для заданного слова `word` возвращает `k` ближайших к нему слов из списка `words` (близость слов измеряется с помощью расстояния Левенштейна)

def task_1_3(word, nums, words):
    words_tuple = list()
    for wor in words:
        words_tuple.append(tuple([edit_distance(word, wor), wor]))
    words_tuple.sort(key = lambda x: x[0])
    
    result = list()
    for wor in words_tuple[:nums]:
        result.append(wor[1])
    print ( result )

word="delicious"
k=4
task_1_3(word, k, words)

### Стемминг, лемматизация

2.1 На основе результатов 1.1 создайте `pd.DataFrame` со столбцами: 
    * word
    * stemmed_word 
    * normalized_word 

Столбец `word` укажите в качестве индекса. 

Для стемминга воспользуйтесь `SnowballStemmer`, для нормализации слов - `WordNetLemmatizer`. Сравните результаты стемминга и лемматизации.

from nltk.stem import SnowballStemmer, WordNetLemmatizer

def task_2_1(wordS):

  wordS['word'] = list(words)
  wordS = wordS.set_index('word')

  stemmer = SnowballStemmer('english')
  lemmatizer = WordNetLemmatizer()
  for index in wordS.index:
      wordS['stemmed_word'][index] = stemmer.stem(index)
      wordS['normalized_word'][index] = lemmatizer.lemmatize(index)

  for index, row in wordS.iterrows():
      print(edit_distance(row[0], row[1]))

wordS = pd.DataFrame(columns=['word', 'stemmed_word', 'normalized_word'])

task_2_1(wordS)

2.2. Удалите стоп-слова из описаний рецептов. Какую долю об общего количества слов составляли стоп-слова? Сравните топ-10 самых часто употребляемых слов до и после удаления стоп-слов.

from nltk.corpus import stopwords
from nltk.probability import FreqDist


def task_2_2(stop):

  buf_recipes = pd.DataFrame(columns=['description', 'filtered_description'])
  buf_recipes['description'] = preprocessed["name"]
  buf_recipes['filtered_description'] = [list() for _ in range(len(buf_recipes['description']))] 

  common_counts = 0
  filtered_counts = 0
  for index, row in buf_recipes.iterrows():
      try:
          for word in word_tokenize(row['description']):
              common_counts += len(word)
              if word.casefold() not in stop:
                  filtered_counts += len(word)
                  buf_recipes['filtered_description'][index].append(word)
          buf_recipes['filtered_description'][index] = " ".join(buf_recipes['filtered_description'][index])
      except TypeError:
          pass


  print(f'percentage:\n {(common_counts - filtered_counts) / common_counts}')
  common_text = ""
  filtered_text = ""
  for index, row in buf_recipes.iterrows():
      try:
          common_text += row['description'] + ' '
          filtered_text += row['filtered_description'] + ' '
      except TypeError:
          continue
  FreqDist(nltk.Text(word_tokenize(common_text))).most_common(10)
  FreqDist(nltk.Text(word_tokenize(filtered_text))).most_common(10)


stop = set(stopwords.words("english"))

task_2_2(stop)

### Векторное представление текста

3.1 Выберите случайным образом 5 рецептов из набора данных. Представьте описание каждого рецепта в виде числового вектора при помощи `TfidfVectorizer`

from nltk import sent_tokenize, word_tokenize
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

def task_3_1(recipes_text):
  recipes = preprocessed.sample(5)
  recipes["description"] = recipes["name"].map(sent_tokenize)
  tv = TfidfVectorizer()
  for index, row in recipes.iterrows():
      recipes_text.append(row['description'][0])
  fit = tv.fit_transform(recipes_text)
  tv = fit.toarray()
  print (tv)

recipes_text = list()
task_3_1(recipes_text)

3.2 Вычислите близость между каждой парой рецептов, используя косинусное расстояние (`scipy.spatial.distance.cosine`) Результаты оформите в виде таблицы `pd.DataFrame`. В качестве названий строк и столбцов используйте названия рецептов.

from scipy.spatial import distance

def task_3_2(cosine_table):


  for i in range(len(random_recipes)):
      for j in range(i, len(random_recipes)):
          cosine_table.iloc[i][j] = 1 - distance.cosine(tv_array[i], tv_array[j])
          cosine_table.iloc[j][i] = 1 - distance.cosine(tv_array[i], tv_array[j])
  print(cosine_table)

cosine_table = pd.DataFrame(index=random_recipes['name'], columns=random_recipes['name'])

task_3_2(cosine_table)

3.3 Какие рецепты являются наиболее похожими? Прокомментируйте результат
