==> Big data analysis/Semester 1/11_dask_array_Фамилия_Имя_Группа.ipynb <==
# Dask Array

Материалы:
* Макрушин С.В. Лекция 11: Dask
* https://docs.dask.org/en/latest/array.html
* JESSE C. DANIEL. Data Science with Python and Dask. 

## Задачи для совместного разбора

1. Создайте массив размерностью 1000 на 300000, заполненный числами из стандартного нормального распределения. Исследуйте основные характеристики полученного массива.

%%time
arr_np = np.random.normal(0,1,size=(1_000,300_000))


%%time
arr_da = da.random.normal(0,1,size=(1_000,300_000), chunks=(1_000,30_000))


arr_da

2. Посчитайте сумму квадратов элементов массива, созданного в задаче 1. Создайте массив `np.array` такого же размера и сравните скорость решения задачи с использование `da.array` и `np.array`

%%time
np.power(arr_np,2).sum()

%%time
da.power(arr_da,2).sum().compute()

3. Визуализируйте граф вычислений для задачи 2.

## Лабораторная работа 11

from google.colab import drive
drive.mount('/content/drive')

import dask.array as da
import dask
import h5py
from functools import reduce
import numpy as np
from tqdm import tqdm # для отображения прогресса

lst=range(10000000)
for i in tqdm(lst): # то, как это делается
  pass

1. Считайте датасет `recipe` из файла `minutes_n_ingredients_full.hdf5` в виде `dask.array`. Укажите аргумент `chunks=(100_000, 3)` при создании массива. Выведите на экран основную информацию о массиве.

# filk= h5py.File("/content/drive/MyDrive/Colab Notebooks/lab 11/minutes_n_ingredients_full.hdf5", 'r')




def load ():
  with h5py.File("/content/drive/MyDrive/Colab Notebooks/lab 11/minutes_n_ingredients_full.hdf5", "r") as f:
    data = f['recipe']
    print(type(data))
  return data
#   #...
  

data=load()

# recipe_da = da(data,chunks=(100_000, 3)).compute()
# recipe_da

f = h5py.File("/content/drive/MyDrive/Colab Notebooks/lab 11/minutes_n_ingredients_full.hdf5", "r")
data = f['recipe']
data = da.from_array(data,chunks=(100_000, 3)).compute()
#....
#data = f['recipe']
print (type(data))
f.close()

temp_file = h5py.File("/content/drive/MyDrive/Colab Notebooks/lab 11/minutes_n_ingredients_full.hdf5", "r" )
temp_dataset = h5py.File("recipes.hdf5", "w")
temp_dataset.create_dataset("recipes", None, None,temp_file['recipe'], chunks=True)
Chunks_Temp_files = h5py.File("recipes.hdf5", "r")


data

2. Вычислите среднее значение по каждому столбцу, кроме первого. 

da.mean(data, axis=0)

res=da.mean(data, axis=0).compute()
a=res[1]
b=res[2]
print (res)

3. Исследуйте, как влияет значение аргумента `chunks` при создании `dask.array` на скорость выполнения операции поиска среднего. 

da.from_array(data)

da.from_array(data,chunks=(100_000, 3))

4. Выберите рецепты, время выполнения которых меньше медианного значения

#ASK

re=data[:,1]
ress=da.from_array(re)
da.median(ress)
#print (da.median(data[:,1]))


ka=data[:,1]
ba=data[:,2]

med=np.median(ka)
ded=np.median(ba)

print (ka[ka < med])
print (ba[ba < ded])
result1=ka[ka < med]
result2=ba[ba < ded]


5. Посчитайте количество каждого из возможных значений кол-ва ингредиентов

# da.unique(data).compute()
# unique, counts = da.unique(data, return_counts=True)
# dict(zip(unique, counts))

#a = вф.array([0, 3, 0, 1, 0, 1, 2, 1, 0, 0, 0, 0, 1, 3, 4])
unique, counts = np.unique(data[:,1:2], return_counts=True)
print (da.asarray((unique, counts)).T)
ree=da.asarray((unique, counts), chunks=(1000,1000)).T.compute()
ree

unique, counts = np.unique(data[:,1:2], return_counts=True)
np.asarray((unique, counts)).T


unique = np.unique(data[:,2])
ree=da.asarray(unique).compute()
print (ree)
print (len(ree), " length")

unique, counts = np.unique(data[:,2], return_counts=True)
print (da.asarray((unique, counts)))
ree=da.asarray((unique, counts)).T.compute()
ree


6. Найдите максимальную продолжительность рецепта. Ограничьте максимальную продолжительность рецептов сверху значением, равному 75% квантилю.

ogran=np.quantile(data[:,1], 0.75)
print ((data[data < ogran]))
re=data[data < ogran]
np.max(re)


7. Создайте массив `dask.array` из 2 чисел, содержащих ваши предпочтения относительно времени выполнения рецепта и кол-ва ингредиентов. Найдите наиболее похожий (в смысле $L_1$) рецепт из имеющихся в датасете.

x = da.random.random((1, 2), chunks=(1, 1)).compute()
#print (x[0])
close_0=1000000000
close_1=1000000000
print (x[0][0] , x[0][1])
id=-100
for i in data:
  if (abs(x[0][0]-i[1])<close_0 and abs(x[0][1]-i[2])<close_1 ):
    close_0=abs(x[0][0]-i[1])
    close_1=abs(x[0][1]-i[2])
    id=i[0]
    print (i[1], i[2])

print (id , close_0, close_1)

x =[[200,43]]
close_0=1000000000
close_1=1000000000
id=0
for i in data:
    if (abs(x[0][0]-i[1])<close_0 and abs(x[0][1]-i[2])<close_1 ):
        close_0=abs(x[0][0]-i[1])
        close_1=abs(x[0][1]-i[2])
        print (i[1] , i[2])
        id=i[0]

print (id)


preferences = da.random.random((1, 2), chunks=(1, 1)).compute()
distance = data().to_dask_dataframe().iloc[:, 1:].sub(preferences)
print ( data().to_dask_dataframe().loc[ distance.apply(np.linalg.norm, axis=1, meta='int').idxmin().compute()].compute() )


8. Работая с исходным файлом в формате `hdf5`, реализуйте алгоритм подсчета среднего значения в блочной форме и вычислите с его помощью среднее значение второго столбца в массиве.

Блочный алгоритм вычислений состоит из двух частей:
1. загрузка фрагмента за фрагментом данных по `blocksize` элементов и проведение вычислений на этим фрагментом;
2. агрегация результатов вычислений на различных фрагментах для получения результата на уровне всего набора данных.

Важно: при работе с `h5py` в память загружаются не все элементы, а только те, которые запрашиваются в данный момент

Lenn = ((Chunks_Temp_files['recipes'][part][:, 1].sum(), Chunks_Temp_files['recipes'][part][:, 1].size) for part in Chunks_Temp_files['recipes'].iter_chunks())
sum_for_minutes = reduce (lambda mas, l: (mas[0] + l[0], mas[1] + l[1]), Lenn)
print ( sum_for_minutes[0] / sum_for_minutes[1])
