==> Big data analysis/Semester 1/12_dask_bag_Фамилия_Имя_Группа.ipynb <==
# Dask Bag

Материалы: 
* Макрушин С.В. Лекция 12: Map-Reduce
* https://docs.dask.org/en/latest/bag.html
* Jesse C. Daniel. Data Science with Python and Dask. 

## Задачи для совместного разбора

1. Считайте файл `Dostoevskiy Fedor. Igrok - BooksCafe.Net.txt` и разбейте на предложения. Подсчитайте длину (в кол-ве символов) каждого предложения.

2. Считайте файл `Dostoevskiy Fedor. Igrok - BooksCafe.Net.txt` и разбейте на предложения. Выведите предложения, длина которых не более 10 символов.

3. На основе списка предложений из задачи 1-2 создайте `dask.bag`. Рассчитайте среднюю длину предложений в тексте.

4. На основе файла `addres_book.json` создайте `dask.bag`. Посчитайте количество мобильных и рабочих телефонов в наборе данных

## Лабораторная работа 12

pip install fsspec

from google.colab import drive
drive.mount('/content/drive')

import dask.bag as db
import json
import re
import pandas as pd

1. В файлах архиве `reviews_full.zip` находятся файлы, содержащие информацию об отзывах к рецептам в формате JSON Lines. Отзывы разделены на файлы в зависимости от оценки (например, в файле `reviews_1.json` находятся отзывы с оценкой 1). Считайте файлы из этого архива в виде `dask.bag`. Преобразуйте текстовое содержимое файлов с помощью модуля `json`. Выведите на экран первые 5 элементов полученного `bag`.

b = db.read_text('/content/reviews_full.zip' ).map(json.loads)
b

#WARNING TEST


# rev= db.read_text('/content/danns/reviews_0.json').take(3)
# result = list(rev)
# result
for i in range(4):
  rev = db.read_text(f'/content/danns/reviews_{i}.json').map(json.loads).take(5)

  result = re.findall(r'\w', f'/content/danns/reviews_{i}.json')
  rating = result[20]
  print (type(rev[0]))
  
  rev[0].update({"rating":rating})
  print (rev[0])

  


def jss (i):
  rev = db.read_text(f'/content/danns/reviews_{i}.json').map(json.loads).take(5)

  result = re.findall(r'\w', f'/content/danns/reviews_{i}.json')
  rating = result[20]
  print (type(rev[0]))
  
  rev[0].update({"rating":rating})
  print (rev[0])


b.to_textfiles("/content/sample_data/*.json")  
#b.to_textfiles("/content/sample_data/*.json", name_function=jss)  

import dask
import json
import os

k = dask.datasets.make_people()                 # Make records of people
k.map(json.dumps).to_textfiles('/content/test/*.json')

b = db.read_text('/content/test/*.json').map(json.loads)
b

b.filter(lambda record: record['age'] > 30).take(2)

#WARNING END OF TEST

2. Модифицируйте функцию разбора JSON таким образом, чтобы в каждый словарь c информацией об отзыве добавить ключ `rating`. Значение получите на основе названия файла (см. аргумент `include_path`), использовав для этого регулярное выражение.

def json_rating(files):
  rating = json.loads(files[0])
  comp=re.compile("reviews_(.*?)\.json")
  rating['rating']= int(re.findall(comp, files[1])[0])
  return rating

output_with_rating = db.read_text(f'/content/drive/MyDrive/Colab Notebooks/lab 12/reviews_*.json', include_path=True).map(json_rating)

3. Посчитайте количество отзывов в исходном датасете.

output_with_rating.count().compute()

4. Отфильтруйте `bag`, сохранив только отзывы, оставленные в 2014 и 2015 годах.

b.filter(lambda record: record['date'] > '2013-31-12' and record['date'] < '2016-01-01').take(5)

RESULT = output_with_rating.filter(lambda record: record['date'] > '2013-31-12' and record['date'] < '2016-01-01')

5. Выполните препроцессинг отзывов:
    * привести строки к нижнему регистру
    * обрезать пробельные символы в начале и конце строки
    * удалите все символы, кроме английских букв и пробелов
    
Примените препроцессинг ко всем записям из `bag`, полученного в задании 4.

def P_O_Y(files):
  pre = re.compile(r'[^a-zA-Z\s]')
  files['review'] = pre.sub('', files['review'].lower().strip())
  return (a)


RESULT.map(P_O_Y).take(10)

6. Посчитайте количество отзывов в датасете, полученном в результате решения задачи 5. В случае ошибок прокомментируйте результат и исправьте функцию препроцессинга.

RESULT.count().compute()

7. Посчитайте, как часто в наборе, полученном в задании 5, встречается та или иная оценка.

#RESULT.distinct(key=lambda x: x['rating']).compute()
TEMP_RES=RESULT.pluck('rating')
TEMP_RES.frequencies().compute()

8. Найдите среднее значение `rating` в наборе, полученном в задании 5.

TEMP_RES.mean().compute()

9. Используя метод `foldby`, подсчитать максимальную длину отзывов в зависимости от оценки `rating` в наборе, полученном в задании 5.

def binop(total, min):
  if len(min['review']) > total:
      return len(min['review'])
  return total

data.foldby('rating', binop,0).compute()
